Share Lynx crosswalk audit helpers
